+++ .github/workflows/pullrequestCI.yml
@@ -6,6 +6,9 @@ jobs:
   build:
 
     runs-on: ubuntu-latest
+    defaults:
+      run: 
+         working-directory: Service/Source
     strategy:
       matrix:
         dotnet-version: ['7.x']
