--- Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using psaltos
-@namespace psaltos.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
