--- Program.cs
@@ -1,17 +1,0 @@
-var builder = WebApplication.CreateBuilder(args);
-
-// Add services to the container.
-
-builder.Services.AddControllers();
-builder.Services.AddSingleton<DapperContext>();
-var app = builder.Build();
-
-// app.UseHttpsRedirection();
-app.UseStaticFiles();
-app.UseRouting();
-
-app.MapControllers();
-
-app.MapFallbackToFile("index.html");
-
-app.Run();
