--- Psaltos.Tests/UnitTest1.cs
@@ -1,15 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace Psaltos.Tests
-{
-    [TestClass]
-    public class UnitTest1
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-            Assert.AreEqual(true, true);
-        }
-    }
-}
-
