--- Psaltos.Tests/Usings.cs
@@ -1,1 +1,0 @@
-
