+++ Service/Source/ClientApp/aspnetcore-react.js
