+++ Service/Source/ClientApp/public/manifest.json
