+++ Service/Source/ClientApp/src/service-worker.js
