+++ Service/Source/ClientApp/src/serviceWorkerRegistration.js
