+++ Service/Source/Models/Asset.cs
