--- package.json
@@ -1,7 +1,0 @@
-{
-  "dependencies": {
-    "aws-sdk": "^2.1323.0",
-    "semantic-ui-css": "^2.5.0",
-    "semantic-ui-react": "^2.1.4"
-  }
-}
